To test: 	run thr.cr
Thread.new allocates a new fiber and attempts to switch to it.

The program crashes when the following happens:
* `fiber.thread_run` is used (in thread.cr) instead of `@ret = @func.call(@arg)`
* `GC.collect` is called the Thread.new (fiber) block (in thr.cr)

The program runs correctly if any of:
* GC.collect is not called within a new thread/fiber pair
* the block passed to Thread.new runs in the thread and not the new Fiber

There are more changes in the scheduler and thread groups that had tons of merge
conflicts.   They won't work until the GC issue is resolved anyway.  I suspect something
is off with stack and/or fiber marking.